% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{addfilter}
\alias{addfilter}
\title{addfilter}
\usage{
addfilter(M, fid, fval, fexcl)
}
\arguments{
\item{M}{is the meta data connection structure}

\item{fid}{item id of the filter}

\item{fval}{filter values for this item}

\item{fexcl}{flag to exclude values}
}
\description{
Add the selected filter item and values to the filter structure 'currfilters'
}
