% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filters}
\alias{filters}
\title{filters}
\usage{
filters(input, output, session, M, D)
}
\arguments{
\item{input}{is shiny input variable}

\item{output}{is shiny output variable}

\item{session}{is shiny session variable}

\item{M}{is the meta data connection structure}

\item{D}{is the data connection structure}
}
\description{
facility to create, save and selected saved filters
}
