% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filtersUI}
\alias{filtersUI}
\title{filtersUI}
\usage{
filtersUI(id, dimgrp, M)
}
\arguments{
\item{id}{is caller id}

\item{dimgrp}{is the dimension group applicable for selected measures}

\item{M}{is the meta data connection structure}
}
\description{
UI for facility to create, save and selected saved filters
}
